+++ tests/headers/struct_with_incomplete_array.h
@@ -1,0 +1,3 @@
+struct incomplete_array {
+    int x; int y[];
+};